refactor(job-handler): rename jobhandling package to job
