+++ connector-runtime/connector-runtime-spring/src/main/java/io/camunda/connector/runtime/outbound/job/JobForLog.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.connector.runtime.outbound.jobhandling;
+package io.camunda.connector.runtime.outbound.job;
 
 import io.camunda.client.api.response.ActivatedJob;
 import java.util.Map;
